Add sources for CRQC following closely behind
One source is the US government. A statement by Biden isn't the greatest
source, and I did add a few more sources to the bibliography, but I
can't really find a definitive source that shows how CRQC is really
closely going to follow quantum practicality, or even that it's likely
to happen at all.

The Randd corporation research actually points in the opposite
direction, with the requisite requirements being well outside our reach
at the moment, and the Ericsson researchers imply that CRQC is out of
reach for the foreseeable future.